requirements.txt
python-docx
textract
requests
times
os